Refine MyBatis cloud auto-config conditions
Switch `MyBatisCloudAutoConfiguration` from `@ConditionalOnMyBatisEnabled` to `@ConditionalOnMyBatisAvailable` so activation is tied to MyBatis availability checks. Add `microsphere-spring-boot-test` for auto-config test support, refactor `MyBatisCloudAutoConfigurationTest` to use `AutoConfigurationTest` with disabled-property and missing-class scenarios, and add a dedicated integration test to verify Spring Cloud feature beans and `FeaturesEndpoint` wiring.
