+++ microsphere-mybatis-spring-cloud/pom.xml
@@ -122,6 +122,13 @@
             <version>${revision}</version>
         </dependency>
 
+        <!-- Microsphere Spring Boot Test-->
+        <dependency>
+            <groupId>io.github.microsphere-projects</groupId>
+            <artifactId>microsphere-spring-boot-test</artifactId>
+            <scope>test</scope>
+        </dependency>
+
         <!-- H2 -->
         <dependency>
             <groupId>com.h2database</groupId>
