+++ microsphere-mybatis-spring-cloud/src/main/java/io/microsphere/mybatis/spring/cloud/autoconfigure/MyBatisCloudAutoConfiguration.java
@@ -17,7 +17,7 @@
 
 package io.microsphere.mybatis.spring.cloud.autoconfigure;
 
-import io.microsphere.mybatis.spring.boot.autoconfigure.condition.ConditionalOnMyBatisEnabled;
+import io.microsphere.mybatis.spring.boot.autoconfigure.condition.ConditionalOnMyBatisAvailable;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.context.annotation.Configuration;
 
@@ -39,7 +39,7 @@ import org.springframework.context.annotation.Configuration;
  * @see org.mybatis.spring.boot.autoconfigure.MybatisAutoConfiguration
  * @since 1.0.0
  */
-@ConditionalOnMyBatisEnabled
+@ConditionalOnMyBatisAvailable
 @AutoConfigureAfter(name = {
         "org.mybatis.spring.boot.autoconfigure.MybatisAutoConfiguration"
 })
