+++ microsphere-mybatis-spring-cloud/src/test/java/io/microsphere/mybatis/spring/cloud/autoconfigure/MyBatisCloudAutoConfigurationTest.java
@@ -17,22 +17,14 @@
 
 package io.microsphere.mybatis.spring.cloud.autoconfigure;
 
-import io.microsphere.mybatis.executor.LoggingExecutorFilter;
-import io.microsphere.mybatis.executor.LoggingExecutorInterceptor;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+
+import io.microsphere.spring.boot.test.AutoConfigurationTest;
+import org.apache.ibatis.session.SqlSession;
+import org.apache.ibatis.session.SqlSessionFactory;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.cloud.client.actuator.FeaturesEndpoint;
-import org.springframework.cloud.client.actuator.HasFeatures;
-import org.springframework.cloud.client.actuator.NamedFeature;
 
-import java.util.List;
-import java.util.Map;
+import java.util.Set;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.NONE;
 
 /**
@@ -42,46 +34,26 @@ import static org.springframework.boot.test.context.SpringBootTest.WebEnvironmen
  * @see MyBatisCloudAutoConfiguration
  * @since 1.0.0
  */
-@SpringBootTest(classes = {
-        LoggingExecutorFilter.class,
-        LoggingExecutorInterceptor.class,
-        MyBatisCloudAutoConfigurationTest.class
-}, webEnvironment = NONE,
-        properties = {
-                "spring.cloud.features.enabled=true",
-                "management.endpoints.web.exposure.include=*",
-        }
+@SpringBootTest(
+        classes = {
+                MyBatisCloudAutoConfigurationTest.class
+        },
+        webEnvironment = NONE
 )
-@EnableAutoConfiguration
-class MyBatisCloudAutoConfigurationTest {
+class MyBatisCloudAutoConfigurationTest extends AutoConfigurationTest<MyBatisCloudAutoConfiguration> {
 
-    @Autowired
-    private Map<String, HasFeatures> hasFeaturesMap;
-
-    @Autowired
-    private FeaturesEndpoint featuresEndpoint;
-
-    @Test
-    public void test() {
-        assertTrue(this.hasFeaturesMap.size() > 0);
-        HasFeatures hadFeatures = this.hasFeaturesMap.get("mybatis.features");
-        assertNotNull(hadFeatures);
-
-        hadFeatures = this.hasFeaturesMap.get("mybatis-spring.features");
-        assertNotNull(hadFeatures);
-
-        hadFeatures = this.hasFeaturesMap.get("mybatis-spring-boot-autoconfigure.features");
-        assertNotNull(hadFeatures);
-
-        hadFeatures = this.hasFeaturesMap.get("microsphere-mybatis-core.features");
-        assertNotNull(hadFeatures);
-
-        assertNotNull(featuresEndpoint.features());
+    @Override
+    protected void configureAutoConfiguredClasses(Set<Class<?>> autoConfiguredClasses) {
     }
 
-    private void assertNamedFeature(List<NamedFeature> namedFeatures, int index, String name, Class<?> type) {
-        NamedFeature namedFeature = namedFeatures.get(index);
-        assertEquals(name, namedFeature.getName());
-        assertEquals(type, namedFeature.getType());
+    @Override
+    protected void configureGlobalDisabledPropertyValues(Set<String> globalDisabledPropertyValues) {
+        globalDisabledPropertyValues.add("microsphere.mybatis.enabled=false");
+    }
+
+    @Override
+    protected void configureGlobalMissingClasses(Set<Class<?>> globalMissingClasses) {
+        globalMissingClasses.add(SqlSession.class);
+        globalMissingClasses.add(SqlSessionFactory.class);
     }
 }
